requirements.txt
openai
langchain
streamlit
python-dotenv
pyPDF2
pandas
langchain-openai

-e .